+++ kret_np_pd/filters.py
@@ -78,137 +78,22 @@ class FilterSampleUtils:
         join_hard_soft: t.Literal["and", "or"] = "and",
     ) -> np.ndarray:
         """
-        Accepts a DataFrame and a boolean filter, and returns a new filter that includes
-        rows that are "nearby" the True rows in the original filter.
+        Return a boolean filter selecting rows "nearby" the True rows in `filt`.
 
-        There are two types of "nearness" that can be applied:
-        1. Hard match: The columns specified in `hard_match` must match exactly for a
-        row to be considered nearby.
+        Nearness has two flavors, combinable via `join_hard_soft`:
+        - `hard_match`: listed columns must exactly equal a seed row's values.
+        - `soft_match`: listed columns must be within a numeric distance of a seed row's
+          value. Pass `{col: dist}` to use `soft_match_default_direction`, or
+          `{col: (dist, direction)}` to override per column. `direction` is one of
+          "both" (|d| <= dist), "forward" (0 <= d <= dist), "backward" (-dist <= d <= 0).
 
-        2. Soft match: The columns specified in `soft_match` must be within a certain distance
-        of the values in the True rows for a row to be considered nearby. The value in
-        `soft_match` is the maximum distance that is allowed, and the direction can be
-        "both", "forward", or "backward". Direction is customizeable per column by passing both
-        (distance, direction).
+        `hard_match_apply` / `soft_match_apply` ('and'/'or') combine conditions within each
+        side; `join_hard_soft` ('and'/'or') combines hard and soft. Returns a length-`len(df)`
+        boolean ndarray.
 
-        The `hard_match_apply` and `soft_match_apply` parameters determine how multiple
-        conditions are combined. If "and", a row must satisfy all conditions to be considered nearby.
-        If "or", a row must satisfy at least one condition to be considered nearby.
-
-        The `join_hard_soft` parameter determines how the hard and soft matches are combined.
-        If "and", a row must satisfy BOTH the hard and soft conditions to be considered nearby.
-        If "or", a row must satisfy EITHER the hard or soft conditions to be considered nearby.
-
-        Examples:
-
-        >>> df = pd.DataFrame({
-        ...     "gameId": [1, 1, 2, 2],
-        ...     "teamId": ["A", "B", "A", "B"],
-        ...     "event":  ["play", "timeout", "play", "play"]
-        ... })
-        >>> # Seed: The timeout in Game 1 for Team B
-        >>> filt = df["event"] == "timeout"
-        >>> # Hard match on both: only find rows that are (Game 1, Team B)
-        >>> get_nearby_rows(df, filt, hard_match=["gameId", "teamId"])
-
-            [False, True, False, False]
-
-        -----------Example 2-----------
-
-        >>> df = pd.DataFrame({
-        ...     "time": [10, 14, 15, 20, 25],
-        ...     "whistle": [False, False, False, True, False]
-        ... })
-        >>> # Seed: The whistle at time 20
-        >>> filt = df["whistle"]
-        >>> # Soft match: 5 units 'backward' (looking at time 15 to 20)
-        >>> get_nearby_rows(
-        ...     df, filt,
-        ...     soft_match={"time": (5, "backward")}
-        ... )
-
-            [False, False, True, True, False]
-
-        -----------Example 3-----------
-
-        >>> df = pd.DataFrame({
-        ...     "time": [10, 11, 12],
-        ...     "dist": [0, 5, 20],
-        ...     "foul": [True, False, False]
-        ... })
-        >>> # Must be within 2 seconds AFTER (forward) AND within 10 distance units (both)
-        >>> get_nearby_rows(
-        ...     df, df["foul"],
-        ...     soft_match={"time": (2, "forward"), "dist": 10},
-        ...     soft_match_apply="and"
-        ... )
-
-        -----------Example 4-----------
-        >>> df = pd.DataFrame({
-        ...     "gameId": [1, 1, 2, 2, 3],
-        ...     "period": [1, 2, 1, 2, 1],
-        ...     "event":  ["goal", "play", "play", "goal", "play"]
-        ... })
-        >>> # Seed: Goals in (Game 1, Period 1) and (Game 2, Period 2)
-        >>> filt = df["event"] == "goal"
-        >>> # Hard match ensures we only get rows within those specific (Game, Period) pairs.
-        >>> # It will NOT match Game 1, Period 2 or Game 2, Period 1.
-        >>> get_nearby_rows(df, filt, hard_match=["gameId", "period"])
-        0     True  # Match (1, 1)
-        1    False  # No match for (1, 2)
-        2    False  # No match for (2, 1)
-        3     True  # Match (2, 2)
-        4    False  # No match for (3, 1)
-
-        -----------Example 5-----------
-        >>> df = pd.DataFrame({
-        ...     "playId": [10, 10, 10, 10, 20, 20],
-        ...     "time":   [1, 4, 5, 6, 5, 6],
-        ...     "dist":   [2, 2, 0, 15, 0, 2],
-        ...     "is_hit": [False, False, True, False, True, False]
-        ... })
-        >>> # Seed: Two hits (one in Play 10 at T=5, one in Play 20 at T=5)
-        >>> filt = df["is_hit"]
-        >>> # 1. hard_match: Stay within the same playId.
-        >>> # 2. soft_match 'time': Overwrite default to look 3s 'backward'.
-        >>> # 3. soft_match 'dist': Use global default 'both' for a 5-unit radius.
-        >>> get_nearby_rows(
-        ...     df, filt,
-        ...     hard_match=["playId"],
-        ...     soft_match_default_direction="both",
-        ...     soft_match={"time": (3, "backward"), "dist": 5},
-        ...     soft_match_apply="and"
-        ... )
-        0    False  # Play 10, T=1 (Out of time window 2-5)
-        1     True  # Play 10, T=4 (In window, In dist)
-        2     True  # Play 10, T=5 (The Seed)
-        3    False  # Play 10, T=6 (Wrong direction: forward)
-        4     True  # Play 20, T=5 (The Seed)
-        5    False  # Play 20, T=6 (Wrong direction: forward)
-
-        -----------Example 6-----------
-        >>> df = pd.DataFrame({
-        ...     "gameId": [1, 1, 2, 2],
-        ...     "time":   [59, 60, 1, 2],  # Game 1 ends at 60, Game 2 starts at 1
-        ...     "event":  [None, "buzzer", None, None]
-        ... })
-        >>> # Seed: The buzzer at the end of Game 1
-        >>> filt = df["event"] == "buzzer"
-        >>> # We want rows within 2 seconds of the buzzer, but ONLY in the same game.
-        >>> get_nearby_rows(
-        ...     df, filt,
-        ...     hard_match=["gameId"],
-        ...     soft_match={"time": 2},
-        ...     join_hard_soft="and"
-        ... )
-        0     True  # Within 2s AND same gameId
-        1     True  # The Seed
-        2    False  # Within 2s (time=1 is close to 60) BUT different gameId
-        3    False  # Different gameId
-        dtype: bool
-
-        >>> # If we changed join_hard_soft to "or", Index 2 would become True
-        >>> # because it satisfies the "nearby time" condition.
+        Full worked examples are appended below at runtime — see
+        `kret_np_pd/_core/get_nearby_rows_examples.py` (or `help(get_nearby_rows)` /
+        `get_nearby_rows?` in Jupyter).
         """
         filt = cls.process_filter(filt, shape=len(df))
         n = len(df)
@@ -256,3 +141,14 @@ class FilterSampleUtils:
             combined = (h & s) if join_hard_soft == "and" else (h | s)
 
         return t.cast(np.ndarray, combined.any(axis=0))
+
+
+# Append the verbose examples onto get_nearby_rows.__doc__. cleandoc strips the source
+# indent of the short docstring so it concatenates flush against the (unindented) examples.
+import inspect as _inspect
+
+from kret_np_pd._core.get_nearby_rows_examples import EXAMPLES_DOC as _EXAMPLES_DOC
+
+FilterSampleUtils.get_nearby_rows.__func__.__doc__ = (
+    _inspect.cleandoc(FilterSampleUtils.get_nearby_rows.__doc__ or "") + "\n\n" + _EXAMPLES_DOC
+)
